requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
twilio>=8.10.0
openai>=1.3.5
python-dotenv>=1.0.0
pydantic>=2.8.0
email-validator>=2.1.0
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
mangum>=0.17.0

